+++ scripts/export_assessment_data.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+import csv
 import sys
 
 sys.path.insert(1, ".")
@@ -7,9 +8,44 @@ import argparse  # noqa: E402
 from distutils.util import strtobool  # noqa: E402
 
 from db.queries.assessment_records.queries import (  # noqa: E402
-    get_assessment_records_by_round_id,
+    get_assessment_records_score_data_by_round_id,
 )
-from scripts.location_utils import export_assessment_data_to_csv  # noqa
+
+
+def export_assessment_data_to_csv(output, filename):
+    """Exports assessment data to a CSV file, splitting the 'Score Date Created'
+    column into separate 'Date Created' and 'Time Created' columns for improved
+    readability.
+
+    Parameters:
+    - output: List of dictionaries representing the assessment data.
+    - filename: Name of the output CSV file.
+
+    """
+    fieldnames = [
+        "Short id",
+        "Application ID",
+        "Score Subcriteria",
+        "Score",
+        "Score Justification",
+        "Date Created",
+        "Time Created",
+    ]
+
+    with open(filename, "w", newline="") as csvfile:
+        writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
+
+        writer.writeheader()
+
+        for row in output:
+            score_date_created = row["Score Date Created"]
+            row["Date Created"] = score_date_created.strftime("%d/%m/%Y")
+            row["Time Created"] = score_date_created.strftime("%H:%M:%S")
+            del row["Score Date Created"]
+
+            writer.writerow(row)
+
+    print(f"Successfully exported to {filename}")
 
 
 def process_assessment_data(round_id, write_csv: bool, csv_location):
@@ -31,7 +67,7 @@ def process_assessment_data(round_id, write_csv: bool, csv_location):
 
     """
 
-    assessment_data = get_assessment_records_by_round_id(round_id)
+    assessment_data = get_assessment_records_score_data_by_round_id(round_id)
     print(f"Extracting requested data for {round_id}")
 
     if write_csv:
